refactor(api): move telegram to modules/ as controller, model, service
